+++ src/training/load_batch_funcs.h
@@ -24,7 +24,7 @@ bool loader_main(string filename, int n_threads, int n_batches, int batch_size, 
                                                                                 vector<vector<float>>& vals,
                                                                                 vector<int>& scores,
                                                                                 vector<float>& results,
-                                                                                int& counter);
+                                                                                uint64_t& counter);
 bool load_from_file(std::ifstream* infile, int n_entries, vector<string>* fen_strs, vector<int>* scores, vector<float>* results);
 inline void create_entry(training_entry* tr, int pos, string fen_str);
 inline uint32_t feature_index(int ksq, int sq, int piece_type, int color);
